+++ nixos/staging.nix
@@ -9,6 +9,7 @@
     ./common.nix
     ./streaming.nix
     ./sync-host-emails.nix
+    ./token-cleanup.nix
     ./nginx.nix
     ./sops.nix
     ./postgresql.nix
@@ -28,6 +29,11 @@
   kpbj.syncHostEmails = {
     enable = true;
     dryRun = true;
+  };
+
+  kpbj.tokenCleanup = {
+    enable = true;
+    dryRun = false;
   };
 
   kpbj.nginx = {